refactor: extract AgentNudgeService from PsiBridgeService
Nudge and message-queue state has no connection to IntelliJ's PSI /
code intelligence layer. It lived in PsiBridgeService only because that
was already the thread-safety boundary between the MCP server thread
and the UI — but that is not a good reason to keep unrelated
responsibilities together.

New service: AgentNudgeService (project-level @Service)
  - pendingNudge, nudgesHeld, messageQueue fields
  - setPendingNudge / consumePendingNudge
  - addOnNudgeConsumed / setOnNudgeConsumed
  - setNudgesHeld
  - setOnNudgeRequested / fireNudge (plugin-initiated nudges)
  - enqueueMessage / removeQueuedMessage / getNextQueuedMessage
  - mergeNudges / appendNudgeToResult static utilities

PsiBridgeService retains only PSI tool-execution responsibilities.
runToolExecution now calls AgentNudgeService.getInstance(project)
to consume and append the nudge to tool results.

Updated callers: CopilotClient, AgentEditSession, ChatToolWindowContent,
PromptOrchestrator. Test class PsiBridgeServiceStaticMethodsTest updated
to reference AgentNudgeService for the moved static methods.

No behaviour change.
